+++ src/App.css
@@ -184,6 +184,16 @@ form {
   width: 100%;
 }
 
+.btn-back {
+  color: black;
+  position: absolute;
+  top: 16px;
+  left: 16px;
+  gap: 0.5rem;
+  font-size: 1.4rem;
+  z-index: 10;
+}
+
 @media screen and (max-width: 764px) {
   .desktop-only {
     display: none;
