+++ src/__tests__/FeatureSection.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import FeatureSection from '../componenets/FeatureSection';
+import FeatureSection from '../components/FeatureSection';
 
 it('renders Footer correctly', () => {
   const tree = renderer.create(<FeatureSection />);
