+++ src/__tests__/Footer.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import renderer, { act } from 'react-test-renderer';
 import { Provider } from 'react-redux';
 import { MemoryRouter } from 'react-router';
-import Footer from '../componenets/Footer';
+import Footer from '../components/Footer';
 import store from '../redux/configureStore';
 
 it('renders Footer correctly', () => {
