+++ src/__tests__/Header.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import renderer, { act } from 'react-test-renderer';
 import { MemoryRouter } from 'react-router';
 import { Provider } from 'react-redux';
-import Header from '../componenets/Header';
+import Header from '../components/Header';
 import store from '../redux/configureStore';
 
 it('renders Footer correctly', () => {
