+++ src/__tests__/Login.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import renderer, { act } from 'react-test-renderer';
 import { MemoryRouter } from 'react-router';
 import { Provider } from 'react-redux';
-import Login from '../componenets/auth/Login';
+import Login from '../components/auth/Login';
 import store from '../redux/configureStore';
 
 it('renders Footer correctly', () => {
