+++ src/pages/Homepage.js
@@ -1,9 +1,9 @@
 import 'react-responsive-carousel/lib/styles/carousel.min.css';
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import Carousel from '../componenets/Carousel';
+import Carousel from '../components/Carousel/Carousel';
 import { fetchHouse } from '../redux/house/house';
-import Navbar from '../componenets/navbar/Navbar';
+import Navbar from '../components/navbar/Navbar';
 
 const Home = () => {
   const dispatch = useDispatch();
