+++ src/pages/Landing/Landing.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import '../Landing.scss';
+
 import { Container } from 'react-bootstrap';
-import Header from '../componenets/Header';
-import FeatureSection from '../componenets/FeatureSection';
-import Footer from '../componenets/Footer';
-import '../landing.css';
+import Header from '../../components/Header';
+import FeatureSection from '../../components/FeatureSection';
+import Footer from '../../components/Footer';
+import './landing.css';
 
 function Landing() {
   return (
