+++ src/pages/Landing/Landing.scss
@@ -1,8 +1,7 @@
 @import url("https://fonts.googleapis.com/css2?family=Lato:wght@100;300;700&display=swap");
 
 .container {
-  max-width: 768px !important;
-  padding: 0 !important;
+  padding: 0;
 }
 
 .slack {
@@ -12,8 +11,7 @@
 .header {
   width: 100%;
   height: 400px;
-  background: url("./assets/3vau5vtfa3qn7k8v.jpg") bottom center no-repeat;
-  padding: 30px;
+  background: url("./assets/background.jpeg") bottom center no-repeat;
 
   &-content {
     width: 215px;
@@ -28,7 +26,6 @@
     &__desc {
       font-weight: 600;
       font-size: 14px;
-      // color: rgb(243, 113, 38);
       color: #000;
       text-shadow: 0 1px 1px rgba($color: #000, $alpha: 0.4);
     }
